% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.W.R
\name{observed.W}
\alias{observed.W}
\title{Calculate observed W statistic of CANOVA}
\usage{
observed.W(x, y, k, tie_shuffle)
}
\arguments{
\item{x}{A vector containing values of a continuous variable (X).}

\item{y}{A vector containing values of a continuous variable (Y).}

\item{k}{the neighborhood structure or bin size of the dataset.}

\item{tie_shuffle}{the number of random shuffle times when there are ties in X.}
}
\value{
observed_W  The value of observed W statistic of CANOVA with or without ties.
}
\description{
observed.W
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
\references{
Wang Y, Li Y, Cao H, et al. Efficient test for nonlinear dependence of two continuous variables[J]. BMC bioinformatics, 2015, 16(1): 1.
}

