% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.W.R
\name{permutation.W}
\alias{permutation.W}
\title{Parallel computing permutations of W statistic of CANOVA}
\usage{
permutation.W(x, y, k, perm, clusters)
}
\arguments{
\item{x}{A vector containing values of a continuous variable (X).}

\item{y}{A vector containing values of a continuous variable (Y).}

\item{k}{the neighborhood structure or bin size of the dataset.}

\item{perm}{the number of permutations to calculate permutation pvalues.}

\item{clusters}{the number of "SOCK" cluster type slave nodes to implement parallel computing on the local machine.}
}
\value{
permutations The permutations of observed W statistic of CANOVA.
}
\description{
permutation.W
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
\references{
Wang Y, Li Y, Cao H, et al. Efficient test for nonlinear dependence of two continuous variables[J]. BMC bioinformatics, 2015, 16(1): 1.
}

