% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W.no.ties.R
\name{W.no.ties}
\alias{W.no.ties}
\title{Calculate observed W statistic of CANOVA without ties}
\usage{
W.no.ties(x, y, k)
}
\arguments{
\item{x}{A vector containing values of a continuous variable (X).}

\item{y}{A vector containing values of a continuous variable (Y).}

\item{k}{the neighborhood structure or bin size of the dataset.}
}
\value{
W  The value of observed W statistic of CANOVA without ties.
}
\description{
W.no.ties
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
\references{
Wang Y, Li Y, Cao H, et al. Efficient test for nonlinear dependence of two continuous variables[J]. BMC bioinformatics, 2015, 16(1): 1.
}

