% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W.ties.R
\name{W.ties}
\alias{W.ties}
\title{Calculate observed W statistic of CANOVA with ties
W.ties}
\usage{
W.ties(x, y, k, tie_shuffle)
}
\arguments{
\item{x}{A vector containing values of a continuous variable (X).}

\item{y}{A vector containing values of a continuous variable (Y).}

\item{k}{the neighborhood structure or bin size of the dataset.}

\item{tie_shuffle}{the number of random shuffle times when there are ties in X.}
}
\value{
W  The value of observed W statistic of CANOVA with ties.
}
\description{
Calculate observed W statistic of CANOVA with ties
W.ties
}
\author{
Yi Li, \email{liyistat@gmail.com}
}
\references{
Wang Y, Li Y, Cao H, et al. Efficient test for nonlinear dependence of two continuous variables[J]. BMC bioinformatics, 2015, 16(1): 1.
}

